Separated navbar in buttons and search bar. Small refactor on the other files
