+++ e-commerce-gaming/src/Components/HomePage.jsx
@@ -1,25 +1,13 @@
 import React from 'react';
 
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-
-
-
 class HomePage extends React.Component {
-
     render() {
-
-        return (
-            <div>
-                Hello
-            </div>
-        )
-
-
-    }
-    
-    
+      return (
+        <div>
+          Home Page
+        </div>
+      );
+    }    
 }
-
 
 export default HomePage;
