+++ e-commerce-gaming/src/Components/Layouts/Footer.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 
 class Footer extends Component {
-    render() {
-        return (
-            <div>
+  render() {
+    return (
+      <div>
 
-            </div>
-        )
-    }
+      </div>
+    );
+  }
 }
 
 export default Footer;
