+++ e-commerce-gaming/src/Components/Layouts/NavbarElements/NavbarButtons.jsx
@@ -1,28 +1,14 @@
 import React from 'react';
 import {
-  Menu,
   AccountBox,
   SupervisorAccount,
   ShoppingCart
 } from '@material-ui/icons';
-import {
-  Button,
-  IconButton
-} from '@material-ui/core';
+import { Button } from '@material-ui/core';
 
 const styles = {
-  root: {
-    flexGrow: 1
-  },
-  title: {
-    flex: 1
-  },
-  hamburguerBtn: {
-    marginLeft: -12,
-    marginRight: 20
-  },
   button: {
-    margin: 5,
+    margin: 10,
     color: '#FFF',
     backgroundColor: '#1D8BF1'
   },
@@ -33,12 +19,8 @@ const styles = {
 
 class NavbarButtons extends React.Component {
   render() {
-    return <div>
-        {/* TODO: Hide Menu and dispay with buttons inside when width is min. (For mobile) 
-          <IconButton color="inherit" aria-label="Menu" style={styles.hamburguerBtn}>
-            <Menu />
-          </IconButton>
-        */}
+    return (
+      <div>
         <Button variant="contained" size="small" color="inherit" style={styles.button}>
           Login
           <AccountBox style={styles.icon} />
@@ -51,7 +33,8 @@ class NavbarButtons extends React.Component {
           Cart
           <ShoppingCart style={styles.icon} />
         </Button>
-      </div>;
+      </div>
+    );
   }
 }
